DLSR-371: refactor batch update script w/shared utils
