+++ utils/__init__.py
@@ -1,0 +1,11 @@
+"""
+Shared utility modules for ftva-mams-data scripts.
+
+Modules
+-------
+filemaker_utils
+    Client initialization, configuration loading, logging setup,
+    and record retrieval helpers for Filemaker.
+spacy_utils
+    NLP helpers using spaCy.
+"""
